moving mine.get into salt/utils/minions.py adding output opion to mine.get runner
